+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Release v0.0.5
+
+### Added
+- Added new method `setNewEnv` (aliased `withNewEnv`) that configures `Process` to start in fresh environment (do not inherit parent environment variables).
+
+### Fixed
+- Correctly handle environment variables in `execv` method (take into account process config).
+
+---
+
 ## Release v0.0.2
 
 - Added new method `execv` that allows to replace current process by executing